requirements.standalone.txt
# 独立版最小依赖
# 核心框架
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# HTTP服务 (可选)
fastapi>=0.100.0
uvicorn>=0.23.0

# 环境变量
python-dotenv>=1.0.0

# HTTP客户端 (用于搜索工具)
httpx>=0.24.0

# Word文档导出 (可选)
python-docx>=1.0.0

# 数据处理 (可选)
pydantic>=2.0.0
